+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- Added checkbox in `Options\Fields\EncrypytedPasswordField` field to confirm if user want to update the value.
+
 ## [1.13.0] - 2025-03-04
 
 ### Added